Move OS specific methods and attributes
Add a BaseOsMixin and a LinuxMixin subclass that holds
the linux specific methods and attributes
